VMI: Split datavolume logic to its own file
